所有机器类使用ParametersAreNonnullByDefault类注解取消单处注解notnull更新接口

+++ src/main/java/cn/qiuye/gtlextend/api/machine/multiblock/IThreadModifierMachine.java
@@ -2,5 +2,5 @@ package cn.qiuye.gtlextend.api.machine.multiblock;
 
 public interface IThreadModifierMachine {
 
-    int getExtendlThread();
+    long getExtendlThread();
 }
